requirements.txt
algoliasearch==3.0.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
distro==1.9.0
idna==3.7
JPype1==1.5.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pdf2json==0.3
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
pycparser==2.22
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tabula-py==2.9.3
tzdata==2024.1
urllib3==2.2.2
